Move case conversion to its own file
