+++ serde_codegen_internals/src/lib.rs
@@ -5,3 +5,5 @@ pub mod attr;
 
 mod ctxt;
 pub use ctxt::Ctxt;
+
+mod case;
